Cleanups

+++ skylib-java/lib-gui-java8/src/main/java/ch/scaille/gui/mvc/properties/FloatProperty.java
@@ -90,7 +90,7 @@ public class FloatProperty extends AbstractTypedProperty<Float> {
 	@Override
 	public void attach() {
 		super.attach();
-		propertySupport.getMain().firePropertyChange(getName(), this, null, Float.valueOf(value));
+		propertySupport.getMain().firePropertyChange(getName(), this, null, value);
 	}
 
 	@Override
